feat(B.A.S.E): add support for the B.A.S.E mod
ore shadow looks really bad but the mod has no real support for blending like Vanilla does and the shadow is used on non stone blocks. So it will look bad if we use our stone shadow. So we have the transparent the layer and run with that for the time being.