NEW: Tesla and more ...
Made guis for (some small things are not yet done):

* Industrial Foregoing
* Powered Thingies
* Tesla
* Tesla Core Lib

The machines blocks of Industrial Foregoing are made out of plastic sheets. So I decided to make the interface look like plastic also to reflect this. And because the interface of that mod depends partly of the other Tesla mods i had to keep that style to them.

If you dont like this style --- well, just dont use it ;)

Cheers
zISKo